quiz_app/serializer.py
from rest_framework import serializers
from .models import Question


class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = ["id", "question", "option_one", "option_two", "option_three", "option_four", "subject"]