requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.3
darkdetect==0.8.0
distlib==0.3.9
dj-database-url==3.0.1
Django==5.1.7
django-allauth==65.11.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
filelock==3.17.0
gunicorn==23.0.0
lxml==6.0.1
mysqlclient==2.2.7
numpy==2.3.3
openpyxl==3.1.5
packaging==24.2
pandas==2.3.3
pillow==11.3.0
platformdirs==4.3.6
psycopg2==2.9.11
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.1
virtualenv==20.29.3
